requirements.txt
fastapi
uvicorn
transformers
huggingface-hub
python-dotenv
torch
